+++ packages/shared/src/UI/modals/ConnectWalletModal/index.tsx
@@ -1,10 +1,86 @@
 import { forwardRef, useState } from 'react'
 import { useAsyncFn } from 'react-use'
-import { useWeb3Connection, useWeb3Others } from '@masknet/web3-hooks-base'
-import { NetworkPluginID, getSiteType, type SingletonModalRefCreator, pluginIDsSettings } from '@masknet/shared-base'
+import { Trans } from 'react-i18next'
+import { DialogContent, Box, Card, Typography, Paper, Link, dialogClasses } from '@mui/material'
+import { delay } from '@masknet/kit'
+import { Icons } from '@masknet/icons'
+import { WalletIcon, useSharedI18N } from '@masknet/shared'
+import {
+    NetworkPluginID,
+    Sniffings,
+    getSiteType,
+    type SingletonModalRefCreator,
+    pluginIDsSettings,
+} from '@masknet/shared-base'
 import { useSingletonModal } from '@masknet/shared-base-ui'
 import type { Web3Helper } from '@masknet/web3-helpers'
-import { ConnectWallet } from './ConnectWallet.js'
+import { ProviderType } from '@masknet/web3-shared-evm'
+import { makeStyles, MaskColorVar, ActionButton, LoadingBase } from '@masknet/theme'
+import { useProviderDescriptor, useNetworkDescriptor, useWeb3Connection, useWeb3Others } from '@masknet/web3-hooks-base'
+import { InjectedDialog } from '../../contexts/index.js'
+
+const useStyles = makeStyles<{ contentBackground?: string }>()((theme, props) => ({
+    dialog: {
+        minHeight: 'auto !important',
+        [`.${dialogClasses.paper}`]: {
+            minHeight: 'unset !important',
+        },
+    },
+    dialogContent: {
+        padding: theme.spacing(2),
+    },
+
+    cardContent: {
+        padding: theme.spacing('22px', '12px'),
+        borderRadius: 8,
+        background: props.contentBackground ?? theme.palette.background.default,
+    },
+    tipContent: {
+        display: 'flex',
+        flexWrap: 'nowrap',
+        alignItems: 'center',
+        marginTop: 12,
+        backgroundColor: MaskColorVar.warningBackground,
+        padding: '13px 12px',
+        borderRadius: 8,
+    },
+    tipContentText: {
+        color: MaskColorVar.warning,
+        fontSize: 13,
+        marginLeft: 8.5,
+    },
+    tipLink: {
+        color: MaskColorVar.warning,
+        textDecoration: 'underline',
+    },
+    connectWith: {
+        fontSize: '14px',
+        color: Sniffings.is_dashboard_page ? '#07101B' : theme.palette.maskColor.dark,
+        fontWeight: 700,
+    },
+    error: {
+        paddingRight: theme.spacing(1),
+    },
+    progressIcon: {
+        fontSize: 14,
+        color: theme.palette.common.white,
+    },
+    progress: {
+        color: theme.palette.common.black,
+    },
+    warningTriangleIcon: {
+        fontSize: 20,
+    },
+    retryButton: {
+        fontSize: 12,
+        backgroundColor: theme.palette.maskColor.publicMain,
+        color: theme.palette.maskColor.white,
+        '&:hover': {
+            backgroundColor: theme.palette.maskColor.publicMain,
+            color: theme.palette.maskColor.white,
+        },
+    },
+}))
 
 export interface ConnectWalletModalOpenProps {
     pluginID: NetworkPluginID
@@ -17,6 +93,8 @@ export type ConnectWalletModalCloseProps = boolean
 export const ConnectWalletModal = forwardRef<
     SingletonModalRefCreator<ConnectWalletModalOpenProps, ConnectWalletModalCloseProps>
 >((props, ref) => {
+    const t = useSharedI18N()
+
     const [pluginID, setPluginID] = useState<NetworkPluginID>()
     const [providerType, setProviderType] = useState<Web3Helper.ProviderTypeAll>()
     const [networkType, setNetworkType] = useState<Web3Helper.NetworkTypeAll>()
@@ -24,7 +102,7 @@ export const ConnectWalletModal = forwardRef<
     const Web3 = useWeb3Connection(pluginID, { providerType })
     const Others = useWeb3Others(pluginID)
 
-    const [connection, onConnect] = useAsyncFn<() => Promise<true>>(async () => {
+    const [{ loading, value: connected, error }, onConnect] = useAsyncFn(async () => {
         if (!networkType || !providerType) throw new Error('Failed to connect to provider.')
 
         const chainId = Others.networkResolver.networkChainId(networkType)
@@ -33,6 +111,7 @@ export const ConnectWalletModal = forwardRef<
         try {
             const account = await Web3.connect({
                 chainId,
+                providerType,
             })
             if (!account) throw new Error('Failed to build connection.')
         } catch (err) {
@@ -49,33 +128,119 @@ export const ConnectWalletModal = forwardRef<
         }
 
         return true
-    }, [networkType, providerType, pluginID, Others, Web3])
+    }, [pluginID, networkType, providerType, Others, Web3])
+
+    const providerDescriptor = useProviderDescriptor(pluginID, providerType)
+    const networkDescriptor = useNetworkDescriptor(pluginID, networkType)
+
+    const { classes } = useStyles({ contentBackground: providerDescriptor?.backgroundGradient })
 
     const [open, dispatch] = useSingletonModal(ref, {
-        onOpen(props) {
+        async onOpen(props) {
             setPluginID(props.pluginID ?? NetworkPluginID.PLUGIN_EVM)
             setProviderType(props.providerType)
             setNetworkType(props.networkType)
 
-            // connect to wallet
-            setTimeout(async () => {
-                await onConnect()
-                dispatch?.close(true)
-            })
+            await delay(300)
+            const rr = await onConnect()
+            console.log(`DEBUG: rr: ${rr}`)
+            dispatch?.close(true)
         },
     })
 
     if (!open) return null
 
+    if (!pluginID || !providerType || !networkType) return null
+
     return (
-        <ConnectWallet
-            pluginID={pluginID}
-            providerType={providerType}
-            networkType={networkType}
-            connection={connection}
-            open
-            onConnect={onConnect}
+        <InjectedDialog
+            title={t.plugin_wallet_dialog_title()}
+            open={open}
             onClose={() => dispatch?.close(false)}
-        />
+            maxWidth="sm"
+            className={classes.dialog}>
+            <DialogContent className={classes.dialogContent}>
+                <Paper elevation={0}>
+                    <Card className={`${classes.cardContent} dashboard-style`} elevation={0}>
+                        <Box display="flex" alignItems="center">
+                            <WalletIcon
+                                size={30}
+                                badgeSize={12}
+                                mainIcon={providerDescriptor?.icon}
+                                badgeIcon={networkDescriptor?.icon}
+                            />
+                            <Box display="flex" flex={1} flexDirection="column" sx={{ marginLeft: 2 }}>
+                                <Typography className={classes.connectWith}>
+                                    {loading
+                                        ? t.plugin_wallet_connect_to()
+                                        : connected
+                                        ? t.plugin_wallet_connected_to()
+                                        : t.plugin_wallet_connect_to()}{' '}
+                                    {Others.providerResolver.providerName(providerType)}
+                                </Typography>
+                                {loading ? (
+                                    <Box display="flex" alignItems="center">
+                                        <LoadingBase
+                                            className={classes.progressIcon}
+                                            size={14}
+                                            sx={{ marginRight: 1 }}
+                                        />
+                                        <Typography variant="body2" className={classes.progress}>
+                                            {t.initializing()}
+                                        </Typography>
+                                    </Box>
+                                ) : null}
+                                {!loading && error ? (
+                                    <Typography className={classes.error} color="red" variant="body2">
+                                        {error.message?.includes('Already processing eth_requestAccounts') ||
+                                        error.message?.includes(
+                                            "Request of type 'wallet_requestPermissions' already pending for origin",
+                                        )
+                                            ? t.plugin_wallet_metamask_error_already_request()
+                                            : error.message ?? 'Something went wrong.'}
+                                    </Typography>
+                                ) : null}
+                            </Box>
+                            {!connected && error ? (
+                                <ActionButton
+                                    loading={loading}
+                                    color="primary"
+                                    onClick={onConnect}
+                                    disabled={loading}
+                                    className={classes.retryButton}>
+                                    {t.plugin_wallet_connect_to_retry()}
+                                </ActionButton>
+                            ) : null}
+                        </Box>
+                    </Card>
+                </Paper>
+                {providerDescriptor?.type === ProviderType.WalletConnect ? null : (
+                    <Card className={classes.tipContent} elevation={0}>
+                        <Icons.WarningTriangle className={classes.warningTriangleIcon} />
+                        <Typography className={classes.tipContentText} variant="body2">
+                            <Trans
+                                i18nKey="plugin_wallet_connect_tip"
+                                components={{
+                                    providerLink: Others.providerResolver.providerHomeLink(providerType) ? (
+                                        <Link
+                                            className={classes.tipLink}
+                                            target="_blank"
+                                            rel="noopener noreferrer"
+                                            href={Others.providerResolver.providerHomeLink(providerType)}
+                                        />
+                                    ) : (
+                                        <span />
+                                    ),
+                                }}
+                                values={{
+                                    providerName: Others.providerResolver.providerName(providerType),
+                                    providerShortenLink: Others.providerResolver.providerShortenLink(providerType),
+                                }}
+                            />
+                        </Typography>
+                    </Card>
+                )}
+            </DialogContent>
+        </InjectedDialog>
     )
 })
